requirements.txt
Pillow
jxlpy
tqdm
numpy
